requirements.txt
aiogram==3.4.1
python-dotenv
aiohttp
pydantic